refactor : filter 관련 변수 및 메소드를 FilterViewModel로 분리
