+++ presentation/src/main/java/com/example/presentation/ui/map/filter/FilterViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.presentation.ui.map.filter
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
+import com.example.presentation.util.MainConstants
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -11,6 +11,11 @@ import javax.inject.Inject
 @HiltViewModel
 class FilterViewModel @Inject constructor() : ViewModel() {
 
+    private val _isFilteredMarker = MutableStateFlow(false)
+    val isFilteredMarker: StateFlow<Boolean> = _isFilteredMarker
+
+    private val filterSet = mutableSetOf<String>()
+
     private val _isKindFilterClicked = MutableStateFlow(false)
     val isKindFilterClicked: StateFlow<Boolean> get() = _isKindFilterClicked
 
@@ -19,6 +24,30 @@ class FilterViewModel @Inject constructor() : ViewModel() {
 
     private val _isSafeFilterClicked = MutableStateFlow(false)
     val isSafeFilterClicked: StateFlow<Boolean> get() = _isSafeFilterClicked
+
+    fun getFilterSet(): Set<String> {
+        return if (filterSet.isEmpty()) setOf(
+            MainConstants.SAFE_STORE,
+            MainConstants.GREAT_STORE,
+            MainConstants.KIND_STORE
+        )
+        else filterSet.toSet()
+    }
+
+    fun updateFilterSet(certificationName: String, isClicked: Boolean) {
+        if (isClicked) {
+            filterSet.add(certificationName)
+        } else {
+            filterSet.remove(certificationName)
+        }
+    }
+
+    fun updateAllFilterUnClicked() {
+        filterSet.clear()
+        _isKindFilterClicked.value = false
+        _isGreatFilterClicked.value = false
+        _isSafeFilterClicked.value = false
+    }
 
     fun updateKindFilterClicked() {
         _isKindFilterClicked.value = _isKindFilterClicked.value.not()
@@ -32,10 +61,7 @@ class FilterViewModel @Inject constructor() : ViewModel() {
         _isSafeFilterClicked.value = _isSafeFilterClicked.value.not()
     }
 
-    fun updateAllFilterUnClicked() {
-        _isKindFilterClicked.value = false
-        _isGreatFilterClicked.value = false
-        _isSafeFilterClicked.value = false
-        Log.d("테스트","${_isKindFilterClicked.value}")
+    fun updateIsFilteredMarker(isFilteredMarker: Boolean) {
+        _isFilteredMarker.value = isFilteredMarker
     }
 }
